+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - When adding Recalbox card to dashboard, pre-fill the yaml required structure (type, entity, title)
 - Add Recalbox Card options to display or not buttons
+- Add Recalbox Card option to display Rom path (false by default)
 
 
 ## v1.2.6 - 28/01/2026